feat(templates): update existing templates' showcase - INNO-1270
