+++ src/systems/ec/implementations/react/packages/campaign-page/stories/Index.jsx
@@ -1,9 +1,8 @@
 /* eslint-disable import/no-extraneous-dependencies */
-import React from 'react';
 import { storiesOf } from '@storybook/react';
-
 import CampaignPageExample from '../examples/Default';
 
-storiesOf('Campaign page', module).add('default', () => (
-  <CampaignPageExample />
-));
+storiesOf('Templates/Campaign page', module).add(
+  'default',
+  CampaignPageExample
+);
