+++ src/systems/ec/specs/templates/campaign-page/docs/code.mdx
@@ -3,7 +3,7 @@ import CampaignPageExample from '@ecl/ec-react-campaign-page/examples/Default'
 
 <Playground
   system="ec"
-  selectedKind="Campaign page"
+  selectedKind="Templates/Campaign page"
   selectedStory="default"
   hideDemo
 >
