+++ src/systems/ec/specs/templates/standard-page/docs/code.mdx
@@ -1,9 +1,11 @@
 import Playground from '@ecl/website-components/Playground'
+import StandardPageExample from '@ecl/ec-react-standard-page/examples/Default'
 
 <Playground
   system="ec"
-  selectedKind="Standard page"
+  selectedKind="Templates/Standard page"
   selectedStory="default"
+  hideDemo
 >
-No preview available, use the Playground or Fullscreen links
+  <StandardPageExample />
 </Playground>
