+++ src/systems/eu/specs/components/card/demo/data--card-event.js
@@ -8,7 +8,7 @@ module.exports = {
     variant: 'standalone',
     label: 'Better regulation',
     href: '/example',
-    level: '1',
+    level: 1,
   },
   description:
     'Transparently designing and evaluating evidence-based EU legislation, backed by citizens views.',
