+++ src/systems/eu/specs/templates/campaign-page/docs/code.mdx
@@ -1,9 +1,11 @@
 import Playground from '@ecl/website-components/Playground'
+import CampaignPageExample from '@ecl/eu-react-campaign-page/examples/Default'
 
 <Playground
   system="eu"
-  selectedKind="Campaign page"
+  selectedKind="Templates/Campaign page"
   selectedStory="default"
+  hideDemo
 >
-No preview available, use the Playground or Fullscreen links
+  <CampaignPageExample />
 </Playground>
